COMP: Update Elastix to more recent patched version to be compatible with ITK-v5.1rc01
A quick patch was applied in a fork of Elastix (related to typed enums), which may not be necessary in the final ITK-v5.1 release.
We should be able to get back to Elastix master or development branch after build errors related to typed enums are sorted out in ITK and/or Elastix.