ENH: Allow editing and saving of registration presets
Built-in elastix registration presets can be copied into the scene, edited, and saved in the scene. When registration is launched then the in-scene configuration data is written to files and passed to Elastix.

In-scene presets can also be saved as user presets, into the user profile folder (each preset in a separate subfolder, containing all necessary information, so that presets can be easily shared with other just by copying the complete folder content).

see https://github.com/lassoan/SlicerElastix/issues/50